requirements.txt
streamlit>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.22.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
# Successfully installed opencv-python-4.11.0.86
opencv-python>=4.5.0
pyswarms>=1.3.0
tqdm>=4.62.0
